+++ app/src/main/kotlin/com/wire/android/ui/home/conversationlist/Mentions.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import com.wire.android.R
 import com.wire.android.ui.common.UserProfileAvatar
-import com.wire.android.ui.main.conversationlist.common.RowItem
+import com.wire.android.ui.home.conversationlist.common.RowItem
 import com.wire.android.ui.home.conversationlist.common.UnreadMentionBadge
 import com.wire.android.ui.main.conversationlist.common.UserLabel
 import com.wire.android.ui.home.conversationlist.common.folderWithElements
