+++ app/src/main/kotlin/com/wire/android/ui/home/conversationlist/common/RowItem.kt
@@ -1,8 +1,10 @@
-package com.wire.android.ui.main.conversationlist.common
+package com.wire.android.ui.home.conversationlist.common
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -18,10 +20,12 @@ fun RowItem(onRowItemClick: () -> Unit, content: @Composable RowScope.() -> Unit
         Row(
             verticalAlignment = Alignment.CenterVertically,
             modifier = Modifier
-                .padding(Dimensions.conversationItemRowPadding)
+                .height(Dimensions.conversationItemRowHeight)
+                .fillMaxWidth()
                 .clickable {
                     onRowItemClick()
                 }
+
         ) {
             content()
         }
